backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.35
alembic>=1.14.0
psycopg[binary]>=3.2.0
pydantic>=2.11.0
pydantic-settings>=2.6.0
pyyaml>=6.0.2
python-multipart>=0.0.12
python-dotenv>=1.0.1
# auth (phase 3)
argon2-cffi>=23.1.0
itsdangerous>=2.2.0
email-validator>=2.2.0
